Added support for Parallax 16-port servo controller board as an output device. Fixed a potential issue allowing signed integers to be used as baud rates for Pololu devices. Fixed a potential issue not capping the PWM output range on RaspberryPi devices. Fixed a potential issue with some joystick variables not being initialized at the beginning of the constructor. Fixed a few potential issues with unintentional type conversions and truncations.

+++ output/pololu.h
@@ -10,7 +10,7 @@ class Pololu : public Motor {
 public:
 	//Constructor
 	Pololu();
-	Pololu(const char *file, int baud);
+	Pololu(const char *file, unsigned int baud);
 
 	//Destructor
 	~Pololu();
@@ -24,7 +24,7 @@ private:
 	RPM::SerialInterface *pololu;
 
 	//Functions
-	double scalePower(double power);
+	unsigned short scalePower(double power);
 };
 
 #endif //POLOLU_H
